+++ junos/resource_bgp_group_test.go
@@ -38,7 +38,7 @@ func TestAccJunosBgpGroup_basic(t *testing.T) {
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
 							"mtu_discovery", "true"),
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
-							"multipath", "true"),
+							"bgp_multipath.#", "1"),
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
 							"remove_private", "true"),
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
@@ -143,6 +143,10 @@ func TestAccJunosBgpGroup_basic(t *testing.T) {
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
 							"advertise_external_conditional", "true"),
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
+							"bgp_multipath.#", "1"),
+						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
+							"bgp_multipath.0.multiple_as", "true"),
+						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
 							"no_advertise_peer_as", "true"),
 						resource.TestCheckResourceAttr("junos_bgp_group.testacc_bgpgroup",
 							"metric_out_igp_offset", "-10"),
@@ -211,16 +215,16 @@ resource junos_policyoptions_policy_statement "testacc_bgpgroup" {
   }
 }
 resource junos_bgp_group "testacc_bgpgroup" {
-  name                     = "testacc_bgpgroup"
-  routing_instance         = junos_routing_instance.testacc_bgpgroup.name
-  advertise_inactive       = true
-  advertise_peer_as        = true
-  as_override              = true
+  name               = "testacc_bgpgroup"
+  routing_instance   = junos_routing_instance.testacc_bgpgroup.name
+  advertise_inactive = true
+  advertise_peer_as  = true
+  as_override        = true
+  bgp_multipath {}
   cluster                  = "192.0.2.3"
   damping                  = true
   log_updown               = true
   mtu_discovery            = true
-  multipath                = true
   remove_private           = true
   passive                  = true
   hold_time                = 30
@@ -313,12 +317,14 @@ resource junos_bgp_group "testacc_bgpgroup" {
   metric_out_igp_delay_med_update = true
   authentication_key              = "password"
   type                            = "internal"
+  bgp_multipath {
+    multiple_as = true
+  }
   graceful_restart {
     restart_time     = 10
     stale_route_time = 10
   }
 }
-
 `
 }
 func testAccJunosBgpGroupConfigUpdate2() string {